+++ source/lambda/custom_resource_us_east_1/index.js
@@ -13,13 +13,15 @@
 
 let fs = require("fs");
 let path = require("path");
-const zipLocal = require("zip-local");
+let AdmZip = require("adm-zip");
 
 
-const AWS = require('aws-sdk');
-const lambda = process.env.METRICS == "true" ? new AWS.Lambda({ customUserAgent: process.env.SOLUTION_IDENTIFIER, region: 'us-east-1' }) : new AWS.Lambda({ region: 'us-east-1' });
-const ssm = process.env.METRICS == "true" ? new AWS.SSM({ customUserAgent: process.env.SOLUTION_IDENTIFIER }) : new AWS.SSM();
-const wafv2 = process.env.METRICS == "true" ? new AWS.WAFV2({ customUserAgent: process.env.SOLUTION_IDENTIFIER, region: 'us-east-1' }) : new AWS.WAFV2({ region: 'us-east-1' });
+const { Lambda } = require("@aws-sdk/client-lambda");
+const { SSM } = require("@aws-sdk/client-ssm");
+const { WAFV2 } = require("@aws-sdk/client-wafv2");
+const lambda = process.env.METRICS == "true" ? new Lambda({ customUserAgent: process.env.SOLUTION_IDENTIFIER, region: 'us-east-1' }) : new Lambda({ region: 'us-east-1' });
+const ssm = process.env.METRICS == "true" ? new SSM({ customUserAgent: process.env.SOLUTION_IDENTIFIER }) : new SSM();
+const wafv2 = process.env.METRICS == "true" ? new WAFV2({ customUserAgent: process.env.SOLUTION_IDENTIFIER, region: 'us-east-1' }) : new WAFV2({ region: 'us-east-1' });
 
 
 exports.handler = async (event, context) => {
@@ -41,13 +43,15 @@ async function createLambdaEdge() {
     const le_path = "./le.js";
     const tmp_le_path = "/tmp/le.js";
     const code_path = path.resolve(le_zip_path)
+    let zip = new AdmZip();
     try {
         //zipping le.js
         console.log("copy " + le_path + " to " + tmp_le_path);
         fs.copyFileSync(le_path, tmp_le_path);
 
         console.log("zipping " + tmp_le_path + " into " + le_zip_path)
-        zipLocal.sync.zip(tmp_le_path).compress().save(le_zip_path);
+        zip.addLocalFile(tmp_le_path);
+        zip.writeZip(le_zip_path);
         console.log("zip created");
         // Creates Edge Lambda
         const params = {
@@ -57,11 +61,11 @@ async function createLambdaEdge() {
             FunctionName: process.env.STACK_NAME + '_Sig4LE', /* required */
             Handler: 'le.handler', /* required */
             Role: process.env.ROLE_ARN, /* required */
-            Runtime: 'nodejs14.x', /* required */
+            Runtime: 'nodejs18.x', /* required */
             Description: 'Sign sign4 requests'
         };
 
-        let result = await lambda.createFunction(params).promise();
+        let result = await lambda.createFunction(params);
         functionArn = result.FunctionArn;
         await publishLEVersion(functionArn);
     } catch (error) {
@@ -86,7 +90,7 @@ async function publishLEVersion(functionArn) {
         while (!isFunctionStateActive) {
             let response = await lambda.getFunctionConfiguration({
                 FunctionName: functionArn
-            }).promise();
+            });
             console.log(`Response from get function configuration ${JSON.stringify(response)}`)
             if (response.State === 'Active' || retry > 10) {
                 isFunctionStateActive = true
@@ -101,7 +105,7 @@ async function publishLEVersion(functionArn) {
             FunctionName: functionArn
         };
 
-        let result = await lambda.publishVersion(params).promise();
+        let result = await lambda.publishVersion(params);
         await saveToSSM(process.env.LAMBDA_VERSION, `${functionArn}:${result.Version}`)
 
     } catch (error) {
@@ -126,7 +130,7 @@ async function createWafRuleGroup() {
             Rules: [],
         };
 
-        let result = await wafv2.createRuleGroup(params).promise();
+        let result = await wafv2.createRuleGroup(params);
         await saveToSSM(process.env.RULE_ID, result.Summary.Id)
 
     } catch (error) {
@@ -148,7 +152,7 @@ async function saveToSSM(paramName, paramValue) {
         Type: 'String',
         Overwrite: true
     };
-    const request = await ssm.putParameter(params).promise();
+    const request = await ssm.putParameter(params);
     return request.Parameter;
 }