+++ source/lambda/export_params/index.js
@@ -10,14 +10,14 @@
  *  OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions    *
  *  and limitations under the License.                                                                                *
  *********************************************************************************************************************/
- const aws = require('aws-sdk');
- const lambda = new aws.Lambda();
+ const { Lambda } = require("@aws-sdk/client-lambda");
+ const lambda = new Lambda();
  
 
 exports.handler = async (event, context) => {
     console.log("event="+JSON.stringify(event));
 
-    for (var record of event.Records) { 
+    for (const record of event.Records) { 
 
         console.log('Stream record: ', JSON.stringify(record));
 
@@ -56,7 +56,7 @@ exports.handler = async (event, context) => {
             }
         };
         console.log("params="+JSON.stringify(params));
-        const result = await lambda.updateFunctionConfiguration(params).promise();
+        const result = await lambda.updateFunctionConfiguration(params);
         console.log(result);
         console.log(`Lambda function ${process.env.SUBMIT_QUERY_FUNCTION} configuration updated`);