+++ source/lambda/generate_token/nodejs/index.js
@@ -11,10 +11,11 @@
  *  and limitations under the License.                                                                                *
  *********************************************************************************************************************/
 
-const aws = require('aws-sdk');
+const { DynamoDBDocument } = require("@aws-sdk/lib-dynamodb");
+const { DynamoDB } = require("@aws-sdk/client-dynamodb");
 const awsSMD = require("aws-secure-media-delivery");
 
-const docClient = process.env.METRICS == "true" ? new aws.DynamoDB.DocumentClient({customUserAgent: process.env.SOLUTION_IDENTIFIER}) : new aws.DynamoDB.DocumentClient();
+const docClient = process.env.METRICS == "true" ? DynamoDBDocument.from(new DynamoDB({ customUserAgent: process.env.SOLUTION_IDENTIFIER })) : DynamoDBDocument.from(new DynamoDB());
 
 const stackName = process.env.STACK_NAME;
 const tableName = process.env.TABLE_NAME;
@@ -31,48 +32,8 @@ secret.initSMClient();
 awsSMD.Token.setDEBUG(true)
 let token = new awsSMD.Token(secret);
 
-exports.handler = async (event, context) => {
-    
-    console.log(JSON.stringify(event))
-    var id;
-    var viewer_attributes = {};
-    var headers = event.headers;
-    var request_querystrings = event.queryStringParameters;
-    var viewer_ip;
-
-    if(event['queryStringParameters'] && event.queryStringParameters['id']){
-        id = event.queryStringParameters['id'];
-        if(!/^\w+$/.test(id) || (id.length > 200)) return response400;
-		delete request_querystrings['id'];
-    } else {
-        return response400;
-    }
-
-    if(headers['cloudfront-viewer-address']){
-        viewer_ip = headers['cloudfront-viewer-address'].substring(0, headers['cloudfront-viewer-address'].lastIndexOf(':'))
-    } else {
-        viewer_ip = event.requestContext.http.sourceIp;
-    }
-
-    const params = {
-        TableName: tableName,
-        Key:{"id": id}
-    };
-
-    const video_metadata = await docClient.get(params).promise();
-    console.log("From DynamoDB:"+JSON.stringify(video_metadata));
-    if(!video_metadata.Item){
-        return {
-        "statusCode": 404,
-        "body": 'No video asset for the given ID'
-        };
-    }
-
-    const endpoint_hostname = video_metadata.Item['endpoint_hostname'];
-    const video_url = video_metadata.Item['url_path'];
-    const token_policy = video_metadata.Item.token_policy;
-
-    if(token_policy['ip']) viewer_attributes['ip'] = viewer_ip;
+function _populate_country_region_city(token_policy, headers) {
+    const viewer_attributes = {};
 
     if(token_policy['co']){
         if(headers['cloudfront-viewer-country']){
@@ -98,6 +59,15 @@ exports.handler = async (event, context) => {
         }
     }
 
+    return viewer_attributes;
+}
+
+function _populate_viewer_attributes(token_policy, viewer_ip, headers, request_querystrings) {
+    let viewer_attributes = _populate_country_region_city(token_policy, headers);
+    if (viewer_attributes.statusCode) return viewer_attributes;
+
+    if(token_policy['ip']) viewer_attributes['ip'] = viewer_ip;
+
     if(token_policy['headers'] && token_policy['headers'].length > 0){
         viewer_attributes['headers'] = headers;
     }
@@ -106,7 +76,52 @@ exports.handler = async (event, context) => {
         viewer_attributes['qs'] = request_querystrings;
     }
 
-	var original_url;
+    return viewer_attributes;
+}
+
+exports.handler = async (event, context) => {
+    
+    console.log(JSON.stringify(event))
+    let id;
+    const headers = event.headers;
+    let request_querystrings = event.queryStringParameters;
+    let viewer_ip;
+
+    if(event['queryStringParameters'] && event.queryStringParameters['id']){
+        id = event.queryStringParameters['id'];
+        if(!/^\w+$/.test(id) || (id.length > 200)) return response400;
+		delete request_querystrings['id'];
+    } else {
+        return response400;
+    }
+
+    if(headers['cloudfront-viewer-address']){
+        viewer_ip = headers['cloudfront-viewer-address'].substring(0, headers['cloudfront-viewer-address'].lastIndexOf(':'))
+    } else {
+        viewer_ip = event.requestContext.http.sourceIp;
+    }
+
+    const params = {
+        TableName: tableName,
+        Key:{"id": id}
+    };
+
+    const video_metadata = await docClient.get(params);
+    console.log("From DynamoDB:"+JSON.stringify(video_metadata));
+    if(!video_metadata.Item){
+        return {
+        "statusCode": 404,
+        "body": 'No video asset for the given ID'
+        };
+    }
+
+    const endpoint_hostname = video_metadata.Item['endpoint_hostname'];
+    const video_url = video_metadata.Item['url_path'];
+    const token_policy = video_metadata.Item.token_policy;
+    const viewer_attributes = _populate_viewer_attributes(token_policy, viewer_ip, headers, request_querystrings);
+    
+
+	let original_url;
 	if(endpoint_hostname && video_url){
 		original_url = endpoint_hostname + video_url;
 	} else {