+++ source/lib/custom_resources/cr_load_athena_config_table.ts
@@ -26,7 +26,7 @@ export class CrLoadSqlParams extends Construct {
   constructor(scope: Construct, id: string, props: IConfigProps) {
     super(scope, id);
 
-    new custom_resources.AwsCustomResource(this, "initDBResource", {
+    new custom_resources.AwsCustomResource(this, "initDBResource", { // NOSONAR
       onCreate: {
         service: "DynamoDB",
         action: "putItem",