+++ source/test/save_auto_session.test.ts
@@ -1,12 +1,12 @@
 const saveAutoSession = require('../lambda/save_auto_session/index.js');
-
-jest.mock("aws-sdk")
+import awsSdkMock from "./__mocks__/aws-sdk-mock";
 
 describe('process.env', () => {
-  const env = process.env
+  const env = process.env;
+  let mocks: any[] = [];
 
   beforeEach(() => {
-      jest.resetModules()
+      mocks = awsSdkMock.mockAllAWSClients();
       process.env = {  
         TTL: "2",
         TABLE_NAME: "myTable"
@@ -14,7 +14,8 @@ describe('process.env', () => {
   })
 
   afterEach(() => {
-      process.env = env
+      process.env = env;
+      awsSdkMock.reseMocks(mocks);
   })
 
   test('Save auto session - non array event', async () => {